Add --public_headers flag
Before, the tool assumed there will be just one header to generate bindings
from. In this CL we add a flag for specifying multiple headers.

Behind the scenes we create a virtual filesystem and in it we create a header
file that includes all headers from --public_headers. We pass this single file
to the Clang tool.

To achieve this we have to use a different ClangTool API than before. The major
side effect of this change is that we no longer have to pass
--use_tool_args_for_compile flag to disable Blaze integration, because there is
no Blaze integration support in this new API :)

PiperOrigin-RevId: 390600337