Handle simple inline functions
This Cl adds support for generating bindings for simple C++ functions. Inline functions may not codegenned in the C++ library, and Clang doesn't know if Rust will need the function or not. Therefore in order to make inline functions callable from Rust we need to generate a C++ file that defines a C++ thunk that delegates to the inline function. Clang will emit code for this thunk and Rust code can call this thunk when needed.

This is not great runtime performance wise in regular builds (inline function will not be inlined, there will always be a function call), but it is correct. ThinLTO builds will be able to see through the thunk and inline code across the language boundary. For non-ThinLTO builds we plan to implement[]

This CL also adds `includes` field to the IR. In order to generate C++ thunk in all the cases clang needs to be able to access declarations from public headers of the C++ library.

PiperOrigin-RevId: 392730011