Add pointer support to C++ codegen.
The IR does not have any special information currently for e.g. the pointer types. In the future, this should probably become some kind of enum. I'm leaving this to a followup CL because it's currently not needed (all information is localized to the two functions `format_rs_type` and `format_cc_type`), and because it's actually a bit annoying to do.

PiperOrigin-RevId: 397255306