Add const.
Note the distinction here between Rust-level constness and C++-level constness: in Rust, const is a property of pointers. In C++, it's a property of types.

And so for example:

| C++                          | Rust                     |
|------------------------------|--------------------------|
| `const T `                   | `T`                      |
| `const T*` (aka `T const *`) | `*const T`               |
| `U<const T>`                 | who knows? Depends on U. |

We preserve the `const` bit in the IR for C++ codegen, even though it has no impact on the Rust after the rust type is determined.

Similarly, in the future, `volatile` is even worse than `const`: it isn't a property of objects, or of pointers, but of **individual accesses**. This is the correct model, but makes it even less like C++.

---

Extra notes:

There are two types for which cv-qualification does not apply: references and functions. So strictly speaking, much as the current structure allows for you to make a nonsensical `* <T, U>` or `*<>`, it also allows for a nonsensical cv-qualified reference type of function type (when we add those). Something to keep in mind if we tighten up the type struct.

PiperOrigin-RevId: 397308355