Don't separate all tokens by spaces in cc_tokens_to_string().
We only emit a space if it is needed to separate an identifier or literal from
another identifier or literal.

The intent of this change is to make it possible to emit include statements for
standard headers, where the angle brackets should not be separated from the name
of the header by spaces. (Otherwise, the preprocessor would interpret the spaces
as being part of the header file name and would not find the header.)

PiperOrigin-RevId: 399614876