Expand on notes about cv-qualified function and reference types.
I wanted to figure out what a cv-qualified function or reference even *means*, and the answer is actually pretty straightforward: "the same thing as the non-cv-qualified version".

So rather than being "nonsensical", it's more that this is two different ways to spell the same type. And we may, rather than forbidding it, want to handle it in the same way as we handle alias normalization more generally.

PiperOrigin-RevId: 399619199