Add derives for `Copy`/`Clone` when appropriate.
(Although I am not sure we should really be deriving Copy, so I've added a note to revisit this later.)

When is "appropriate"? Per earlier discussion (e.g. in https://chat.google.com/room/AAAAl3r59xQ/slCadXgyw20), we limit only to trivially-relocatable types, which we identify as exactly those which are "trivial for calls":

> Passing through registers is almost definitionally relocation in the same sense: after a call, it will be as if the object was memcpy'd to a new location. Passing by value, for such an object, is a move constructor call and then a (c++-move-free) relocation operation.

(If a type is not trivial for calls, it's not going to be usable in Rust by value at all, pending work in b/200067242.)

If it *is* trivial for calls, and it has a trivial public (non-deleted) copy constructor, then we derive `Clone` and `Copy`.

PiperOrigin-RevId: 401210466