Add blank lines to the generated bindings files and readjust the order, so that they are more readable.
This uses the approach to add newline characters manually. This produces the
nicest output, but it is also the most annoying.

Alternatives would be to not care about spacing at all, or to use
rustfmt/clang-format to add blank lines indiscriminately, which may be
compromise.

In the end I prioritized user experience.

PiperOrigin-RevId: 401698146