Copy free top-level comments from headers to Rust bindings.
Comments are not traversed during normal AST visitations, so this keeps track of
all comments in the current file and emits IR for all preceding comments
whenever we visit a decl.

This uses source code locations to determine comment order. It also skips
comments that are fully contained within a decl, i.e., comments in the body of
inline functions.

PiperOrigin-RevId: 402777124