Expose IR-from-C++ in ir_testing.rs, and use it to generate the top-level prototypal items.
This has the direct benefit of using "real" data instead of fake data -- for example, the Record in ir_testing.rs was fundamentally impossible (it had a zero size, which no record is allowed to have). Further, since this now uses the C++ source code as an intermediate representation, these will not go out of date as the IR continues to evolve, except for the parts they touch (in particular, the way that they change the names).

I think a future change should likely make these one-time initialization (via e.g. lazy statics) + clone and mutate. I also suspect a similar thing could be done to for function params, and other bits of the IR, but Func and Record are the pieces of IR that change most frequently and are the most complex, so those are the ones that really benefit.

This also, by exporting the ir_from_cc function, allows for tests to do the same trick, and specify their inputs as C++ instead of IR. This reduces the fragility concerns I have around writing non-golden tests, since the C++ source code forms a stable base to test on, rather than the IR which we continue to evolve.

(That latter part is the true motivation here, as discussed in beanbags VC -- in order to make the tests for codegen less fragile and potentially more useful, I want to specify their test data as C++, not as IR literals.)

PiperOrigin-RevId: 404463534