Categorize two types of nontrivial special member functions: those which are nontrivial due **only** to member variables, and the rest.
Taking destructors as an example: the problem is that if we invoke the C++ destructor, it will also destroy all the members. So we have two choices: put every member (in Rust) in a ManuallyDrop, or -- if we know that the C++ destructor doesn't do anything except run all of the members' destructors -- not call the C++ destructor at all, and let Rust invoke the destructors of the members.

Discussions:

* **Idea:** https://chat.google.com/room/AAAA3-o4hPQ/Rc4FFFCdMCw
* **Implementation:** https://chat.google.com/room/AAAAb6i7WDQ/rcRU8T2SaAM

PiperOrigin-RevId: 404774733