Add DeclId to items
Since our IR goes through the JSON
serialization/deserialization at the moment, we need a way to restore graph
edges that don't follow the JSON tree structure (for example between types and
records). We will use DeclId for this.

We put decl_id on RsType and CcType and not onto the MappedType because we expect that those 2 types can reference different Decls later on (for example if we decide to map C++ Status onto Rust Result).

PiperOrigin-RevId: 409359971