Read type identifier from the item in src_code_gen
Before, for Tag types, we were putting the identifier into the Type instance itself. Since we are adding the concept of IR Graph into the SrcCodeGen, we can actually find the associated Item for each relevant type, and retrieve the identifier from there.

This cl modifies the logic so we do that.

A followup to this CL will refactor how types are created in C++ and how are they represented in Rust, very likely to become enums with values like "user_defined" and "primitive" (like int and pointer).

PiperOrigin-RevId: 412817557