Rename Record::decl_id to id, Func::decl_id to record_decl_id.
The intention is to meke it clearer which items have a decl_id (record), and which items reference other items (func, mapped types).

While doing this I realized Func had a decl_id similarly to record, but I believe the intention was for methods to be able to reference their owning record (at least I think Devin is planning to rely on that AFAIK). So I fixed that. It is possible that in the future we will add Func::id field as well.

PiperOrigin-RevId: 413751180