Add lifetimes to IR.
This only changes the IR data structures and fixes any tests that break. It does
not yet add lifetime-specific tests; this will come in a followup CL that
actually populates lifetimes and uses them in code generation. I'm choosing to
do that in a separate CL as this CL is already large enough as it is.

PiperOrigin-RevId: 413901182