Introduce assert_cc_matches/assert_rs_match/assert_cc_not_matches/assert_rs_not_matches
With these it will be easier to write focused unit tests for the src_code_gen logic. Currently these methods support:

* subtree match
* ... wildcard that ignores the entire subtree rooted at a TokenTree (`quote!{fn foo(...)}` for example)

The methods do not yet support mixing wildcard with sibling tokens (`quote!{fn
foo(first: u32, ...)` is not supported for example).

Added usability feature of this solution is the stacked error message on
mismatch and rustfmtted input.

Example usage can be seen in unknown commit.

PiperOrigin-RevId: 415247825