Elide "-> ()" in the generated function definitions.
This CL elides "-> ()" in the generated function definitions (e.g. when
wrapping a C++ function that returns `void`).  This addresses a TODO
from `generate_func` in src_code_gen.rs

The tweaks in `generate_func` are based on what gets returned by the
new helper function that this CL adds to ir.rs:

    impl RsType { pub fn is_unit_type(&self) -> bool { ... } }

PiperOrigin-RevId: 416588628