Major change: additional requirements for rust-reference-safety/Unpin, which now includes "tail padding is safe to write to".
Before, it was assumed a type could be put in a `&mut` if the type was trivially relocatable. However, some types are trivially relocatable, but nonetheless not actually safe to use in Rust via a mutable reference. They can be written to by `memcpy`, but **Rust will `memcpy` an incorrect number of bytes**. These types can still be handled by value, but all references Rust receives from C++ must be `Pin`.

In addition to ensuring types are trivially relocatable, to get a completely idiomatic interface in Rust, those types must also be final leaf types.

This is described more completely in the newly-added doc in google3/devtools/rust/cc_interop/g3doc/unpin.

---

Thanks:

* dmitrig@ for helping organize my thoughts on this, and suggesting `final` and giving up on `[[no_unique_address]]`, instead of my thought of e.g. "`data size == stride`")
* lexer@ and the rest of the Rust room for discussing this at length with me a month or so ago and giving me a lot of intellectual food to chew on.

PiperOrigin-RevId: 417820502