Add an end-to-end test for consuming simple Abseil types and functions.
The test is deactivated as it doesn't actually work yet, but the I think it's a
useful example of the relatively simple cases that we want to get to work. The
idea is that the test can be activated locally to see how far we currently get
as we work towards enabling everything we need for the test.

PiperOrigin-RevId: 419574129