Add a `source_loc` field to `Func`.
This will be needed to generate UnsupportedItems for overloaded functions.

As overloaded functions are already supported by the code generator in specific
cases (namely, constructors taking a single argument), we cannot make the
decision whether an overloaded function is unsupported or not in AstVisitor, so
we need to propagate the source location to the code generator so that it can
generate an UnsupportedItem with source location information.

PiperOrigin-RevId: 420223906