Fix handling of constructors with elided lifetimes.
Before this CL, the generated Default::default and/or From<T>::from
functions would always pass the MaybeUninit<T> `tmp` into the
constructor-wrapping thunk as follows: `tmp.as_mut_ptr()`.  This
wouldn't work when eliding lifetimes, when the thunk parameter type
is `&'a mut SomeStruct` rather than `*mut SomeStruct`.

To fix the problem above, this CL starts passing `tmp` as `&mut tmp`.
This works both with references and with pointers, as long as the type
of `__this` is spelled as: `&'a mut MaybeUninint<SomeStruct>`.

PiperOrigin-RevId: 421065397