Use `format_as_self_param...` for destructors and some constructors.
This function avoids hardcoding `&self` or `&mut self` when generating
code for `Drop::drop` or `Clone::clone`.  This is done by tweaking
`generate_func` to use `format_as_self_param_for_instance_method` not
just for instance methods, but also for constructors and destructors.

As discussed earlier in chat, this CL tweaks `format_as_self_param...`
so that it keeps generating `&mut self` for destructors, even if they
are not annotated with lifetimes.

This CL is required to avoid breaking tests during a follow-up
unknown commit (deduplicating + unifying ctors/dtors/functions handling).

PiperOrigin-RevId: 421916036