Account for textual headers when determining the owning target of a decl
Before this cl, we did not generate bindings for decls coming from textual headers, as they didn't belong to any target/crate. Now, when we inspect a decl coming from a textual header, we go up the #include stack to locate the first header that has an owning target, and that target becomes the owner of the decl.

PiperOrigin-RevId: 423281830