Hardcode "cleanups" for some google3 targets
I stumbled upon this situation:

* //base has base/callback.h, base/callback-specializations.h and some others in hdrs.
* //base depends on //base:callback
* //base:callback has base/callback.h in hdrs, base/callback-specializations.h and those others in srcs.

So we have 2 targets that provide the same header, and one depends on the
another. The interop tooling assumes each header belongs to a single target.
I chatted with pefoley@ and it turns out the public headers in //base are going
to be removed, and the world should depend on
//base:callback to get them. The LSC has not yet started.

To make things work in Rust I'm hardcoding the "fix" which is pretending that
//base doesn't provide those callback headers.

The same situation happens with base/file_toc.h.

PiperOrigin-RevId: 424271720