[rust-interop] Add a temporary way to add lifetime annotations to functions.
This exploits the clang::annotate attribute to attach a function lifetime
annotation to the function declaration. Eventually, this should be replaced by
a more convenient syntax, but in the meantime this allows to start working with
lifetime annotations.

A future CL will add similar temporary support for lifetime parameters on classes/structs.

PiperOrigin-RevId: 424328341