Manually traverse the AST instead of using an `AstVisitor`.
This is in preparation of more precisely determining which types can
be imported or not (potentially because of dependencies), but it does
not yet make that change.

The intent is to not yet change which decls get visited. It also
doesn't apply the lookup mechanism to types, yet. These will be
follow-up changes.

PiperOrigin-RevId: 424577249