Add base class subobjects to the Rust struct.
This is documented in a new markdown file, along with the existing layout tweaks for empty classes.

The tl;dr is: use unaligned bytes to represent base classes, and fix the alignment after the fact using `#[repr(align(N))]`. The base classes are not exposed as struct subobjects, because this would not be correct if they have tail padding: they really must be unaligned bytes.

PiperOrigin-RevId: 426021613