Base class conversion / upcasting support in the IR.
This is pretty much just the IR parts, with a bare minimum of extra Rust bindings codegen to test the IR in an end-to-end fashion. The final implementation of upcasting must also allow for:

* `*mut T`, `*const T` pointers.
* `&T`, as this CL already does.
* `Pin<&mut T>` (and probably `Pin<&T>` just for completeness.)

I believe the correct final design is a trait which casts raw pointers, and a blanket impl for the other pointer/reference types as necessary. That'll be done with a supporting trait / library in a followup CL.

BTW, debugging this was extremely gnarly 🙃 See: https://chat.google.com/room/AAAAouY3z6s/j-Pxwq88M6w

TODO in a followup CL: perhaps this should use [AsRef](https://doc.rust-lang.org/std/convert/trait.AsRef.html) etc.?

TODO in a followup CL: should we use `Deref` for the primary base class, or something?

TODO in a followup CL in a different bug: automatically implement methods on the derived class that forward to the base...?

Lots and lots of things that can be done to improve ergonomics, of course!

PiperOrigin-RevId: 426021695