Separate include sections by a newline. See:[]
Without this, `hg fix` (or `g4 fix`) is a pain whenever I end up touching this part of the file -- it reformats the file, and breaks the goldens.

The change is hacky and not really perfect: we happen to know that both sections are definitely not empty, and so can add the whitespace unconditionally. If we ever get any possibly-empty sections, this will start being annoying!

PiperOrigin-RevId: 426063270