Add layout support for C++20 `[[no_unique_address]]`, without making the variables usable.
This is the main thing I want to do, even without a direct need -- just getting the obvious TODO out of the way, while it's still obvious to me how to do the layout.

A followup CL (not yet started) will provide read support so that this can also be tested at runtime, and can be at least slightly more useful.

PiperOrigin-RevId: 426084219