Generate bindings for C++ standard library
This cl implements a new bindings_for_toolchain_headers rule that generates bindings for the C++ standard library headers. The list of headers that we consider public is located under devtools/rust/cc_interop/rs_bindings_from_cc/BUILD. I'll expand it in a followup cl.

Currently the grte headers are considered to be non-public headers that are part of the C++ standard library. We'll create a separate target for them in a followup cl.

We will not be generating bindings for the clang builtin headers.

PiperOrigin-RevId: 426237240