Very first pass: add `Pin` to mut references to `!Unpin` types, including `self`.
non-`mut` references are left alone, as is idiomatic in Rust. (In particular, non-mut references can be freely moved out of pins to begin with![]

I originally tried to avoid handling `self` in this "very first pass", but in fact some tests already depend on being able to call a method on an `!Unpin` struct.

In order to not break it when it invokes inline functions, I added `self` too.

(`Drop` is handled specially, just as it already was for raw pointers, so as to avoid shaking up the code too much.)

This does not:

* allow safely returning or passing them by value, or change the (currently incorrect or nonexistent) FFI for it.
* Specially handle traits in any way (except Drop).
* work for any other pointer type -- once we add supportfor e.g. `unique_ptr`, or Box, or whatever, we'd need to separately handle "should this be pinned?" for smart pointers as well.
* Add a `use` declaration for `Pin` :'(

Important design considerations:

* I put this in `src_code_gen.rs`, rather than `importer.cc`, because it's going to need to integrate closely with the rest of serialization -- I had originally left that for the future, but you can see how it plays out already now with `self` and `Drop`. It's not enough to just silently inject `Pin` everywhere! I think the choice here is correct, but I do need to brign it up because it is different from `Option`.

PiperOrigin-RevId: 430277528