Change syntax for temporary `lifetimes` annotation.
Put the lifetimes in a separate attribute argument instead of doing our own
string parsing, as this is the way the `annotate` attribute was intended to be
used.

A follow-up CL will make similar changes to the uses of the `annotate`
attribute introduced in https://github.com/google/crubit/commit/3c3d574f26e5156d156e861d0779fbb3f731adea.

PiperOrigin-RevId: 430412836