Merge destructor logic with the rest of the non-constructor functions, by forcing it to always be `safe`.
This means that there's now exactly two cases for body generation: "this is the constructor", or, "this is not the constructor". That sets a good stage for adding a third one: "this is the constructor, but !Unpin", or thereabouts.

(I expect we'll refine it even more beyond that, because I haven't yet even thought about return-by-value for !Unpin, which is again similar.)

PiperOrigin-RevId: 430578614