Add license header to golden files
We're open sourcing the interop tooling and for that we need to move the code to
third_party. This CL makes the golden files compliant to the rules there.

This CL modifies the update.sh script, the testing script, and also updates all the golden files using the new update.sh script.

PiperOrigin-RevId: 433435626