Remove src_code_gen_test.cc
This test is not adding much. Compared to src_code_gen.rs tests it also covers
that C++ code correctly reads the strings passed from Rust (FFIIntegration).
There is only a small step from there to a full integration tests such as
goldens. Therefore I think FFI integration is sufficiently tested and
the understandability and focus of the FFIIntegration test is not enough to pay
for the ongoing maintenance (the test needs to be changed each time there is a
change in the generated code).

PiperOrigin-RevId: 434412850