Do not generate bindings for items in namespaces
In order to be able to import items in namespaces we need to add Ids to Items that currently don't have an Id (such as unsupported item). It turns out that that causes determinism check failures ([]

This CL temporarily disables importing items in namespaces, so we can add the Ids, and with them it will be easier to split ImportDecl into more granular pieces with the appropriate logic.

PiperOrigin-RevId: 435031204