Split `Reconstruct` trait into two: a safe trait that's unsafe to call, and an unsafe trait that's safe to call.
(That's maybe a mouthful, but the idea is that one trait has the caller guarantee correctness, and the other has the implementor guarantee it.)

This resolves concerns around base classes: non-final classes will not implement the trait that is safe to call, because it is UB to reconstruct a base class subobject.

This also pretty solidly puts `reconstruct` in the territory of "(perhaps) useful, but not the thing you use by default", at least not unless we can get better UB-guarantees from clang. (For example, if we defined behaviour for all non-polymorphic base classes, we're getting closer to general utility!)

Huge thanks to both richardsmith@ and lexer@ for pointing out this particular usability issue, which I didn't appreciate at the time of writing.

PiperOrigin-RevId: 435668831