Build the dependencies of our bindings in `target` configuration mode
Currently we build these dependencies in `exec` mode, which can confuse the Rust compiler when we build a Rust target that depends on bindings and also depends on one of these dependencies via a regular path.

We fix this by having the bindings dependencies be attached to the aspect instead of the bindings generator tool (which is built in `exec` mode). As we attached the aspect on every `rust_*` rule, we need to `transition` our way out of the dependency cycle that happens when we need to build a bindings dependency target.

PiperOrigin-RevId: 435739419