Rename DeclId to ItemId and make all our items have an item id.
Namespaces will contain an List<ItemId> that we'll use to generate the right items to the right place. For this, Comments and UnsupportedItems also need an id. For comments we use the address of the RawComment as an id, and the UnsupportedItem gets the id of its own decl (as of https://github.com/google/crubit/commit/593559136eb262fe3870c9827a00f232439a7587 we are generating one pre-IR item per decl, so each UnsupportedItem gets a unique id)

PiperOrigin-RevId: 436748182