Move `is_unsafe` into `ImplKind`.
This is not quite the same as the previous "correct by construction" things -- an `ImplKind::Trait` that `is_unsafe` should not exist,
so this invariant should be checked at the time the `ImplKind` is made. In particular, this is true because it really is a property of
the trait as to whether it's unsafe, _not_ a property of the parameters. A future change might actually allow for unsafe trait methods.

Along these lines, I've removed the old check. Every branch creating traits already does the appropriate check, so this is safe.

PiperOrigin-RevId: 437977983