Move `forward_declare` into crubit so it can be used for bindings.
I think, given recent conversations around template instantiations, that we'll want to modify this a bit further --
the `forward_declaration!` and `symbol!` macros are fine as is, but rather than specifically an `IncompleteCast`
trait, I think we will want to change this to use a universal cast, for all types which are the same in C++ but
must be different in Rust. Perhaps that's as simple as renaming the trait.

PiperOrigin-RevId: 438512849