Precompute the entire mapping of lifetime-id to lifetime string.
This simplifies the API: callers don't need to manually construct the mapping when invoking formatting functions, but can just pass the IR as normal, which already has it.

There's also a later goal I want to accomplish: I think we might be able to remove a category of error. it should be possible to guarantee, once an `RsTypeKind` is constructed, that it will format correctly and doesn't have dangling references to lifetimes or the like.

As an intermediate step to that, we would need to move the creation of the mapping earlier, so that it's available when the RsTypeKind is constructed. This is that first step. I also think it's straight up an improvement by itself, as noted above.

PiperOrigin-RevId: 438791470