Move lifetime lookup failures to RsTypeKind-construction-time.
This makes formatting an unfailable operation. Hooray! No more complicated object invariants that lead to formatting failures. :)

In principle, now, for example, we could straight up implement `ToTokens` for RsTypeKind, instead of having a format method.
This requires further work to eliminate the `ir` parameter that might not be worth it, though.

The main point is that errors now occur closer to the real problem, and do not happen during plain formatting.

PiperOrigin-RevId: 438792724