Move crate name evaluation to `RsTypeKind`-creation time.
This makes the `RsTypeKind` completely-represent the type: it does not need any supplementary information from the IR in order to decide what it looks like when written to a Rust token stream.

I think that this is the correct choice from a layering perspective, as it means that we don't need to look "backwards" to earlier layers.

(And, as a self-contained representation, this also means it can straight up implement ToTokens. In other words, where we currently use token streams for types, we can change this to exclusively use structured data in the form of the `RsTypeKind`. I'm doing this in a followup CL.)

PiperOrigin-RevId: 439624523