Store an `RsTypeKind` instead of a `TokenStream` in `CtorNew`.
The former is now just a structured version of the latter, so better to preserve structure. For example, this makes it potentially easier later for CtorNew to have tuples of arguments, as we don't need to reconstruct the RsTypeKind in the caller, but can use whatever is returned as the args for CtorNew.

Discussed a bit in chat (https://chat.google.com/room/AAAAImO--WA/tSx3n_xLbSs), seems uncontroversial. :)))

PiperOrigin-RevId: 439912565