Delete `param_types`, replacing with `param_type_kinds` (which is then renamed to `param_types`.)
Since `RsTypeKind` is now formattable, and equivalently useful to a plain `TokenStream`, we don't need (or want) two separate vectors, and can just use the one.

Most of the work for this was done, accidentally, in unknown commit -- in that CL, I avoided passing in the old `type_kinds` just to keep the parameter count down. But as a result, there's basically no remaining complex use of `type_kinds`, and it can be safely removed.

PiperOrigin-RevId: 439914011